# openapi.json
{
    "openapi": "3.0.0",
    "info": {
      "title": "Star Wars Films API",
      "version": "1.0.0",
      "description": "API for managing Star Wars films"
    },
    "servers": [
      {
        "url": "http://localhost:3000/"
      }
    ],
...
    "paths": {
...
      "/starships": {
        "get": {
          "tags": [
            "starships"
          ],
          "summary": "Get all starships",
          "responses": {
            "200": {
              "description": "Successful response",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/Starship"
                    }
                  }
                }
              }
            }
          }
        },
        "post": {
          "tags": [
            "starships"
          ],
          "summary": "Create a new starship",
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StarshipInput"
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Successful response",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Starship"
                  }
                }
              }
            }
          }
        }
      },
      "/starships/{id}": {
        "get": {
          "tags": [
            "starships"
          ],
          "summary": "Get a starship by ID",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "string"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Successful response",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Starship"
                  }
                }
              }
            }
          }
        },
        "put": {
          "tags": [
            "starships"
          ],
          "summary": "Update a starship by ID",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "string"
              }
            }
          ],
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StarshipInput"
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Successful response",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Starship"
                  }
                }
              }
            }
          }
        },
        "delete": {
          "tags": [
            "starships"
          ],
          "summary": "Delete a starship by ID",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "string"
              }
            }
          ],
          "responses": {
            "204": {
              "description": "Successful response"
            }
          }
        }
      },
      "/transports": {
        "get": {
          "tags": [
            "transports"
          ],
          "summary": "Get all transports",
          "responses": {
            "200": {
              "description": "Successful response",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/Transport"
                    }
                  }
                }
              }
            }
          }
        },
        "post": {
          "tags": [
            "transports"
          ],
          "summary": "Create a new transport",
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransportInput"
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Successful response",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Transport"
                  }
                }
              }
            }
          }
        }
      },
      "/transports/{id}": {
        "get": {
          "tags": [
            "transports"
          ],
          "summary": "Get a transport by ID",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "string"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Successful response",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Transport"
                  }
                }
              }
            }
          }
        },
        "put": {
          "tags": [
            "transports"
          ],
          "summary": "Update a transport by ID",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "string"
              }
            }
          ],
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransportInput"
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Successful response",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Transport"
                  }
                }
              }
            }
          }
        },
        "delete": {
          "tags": [
            "transports"
          ],
          "summary": "Delete a transport by ID",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "string"
              }
            }
          ],
          "responses": {
            "204": {
              "description": "Successful response"
            }
          }
        }
      },
      "/vehicles": {
        "get": {
          "tags": [
            "vehicles"
          ],
          "summary": "Get all vehicles",
          "responses": {
            "200": {
              "description": "Successful response",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/Vehicle"
                    }
                  }
                }
              }
            }
          }
        },
        "post": {
          "tags": [
            "vehicles"
          ],
          "summary": "Create a new vehicle",
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VehicleInput"
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Successful response",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Vehicle"
                  }
                }
              }
            }
          }
        }
      },
      "/vehicles/{id}": {
        "get": {
          "tags": [
            "vehicles"
          ],
          "summary": "Get a vehicle by ID",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "string"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Successful response",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Vehicle"
                  }
                }
              }
            }
          }
        },
        "put": {
          "tags": [
            "vehicles"
          ],
          "summary": "Update a vehicle by ID",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "string"
              }
            }
          ],
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VehicleInput"
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Successful response",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Vehicle"
                  }
                }
              }
            }
          }
        },
        "delete": {
          "tags": [
            "vehicles"
          ],
          "summary": "Delete a vehicle by ID",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "string"
              }
            }
          ],
          "responses": {
            "204": {
              "description": "Successful response"
            }
          }
        }
      },
      "/users": {
        "post": {
          "tags": [
            "users"
          ],
          "summary": "Create a new user",
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserInput"
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Successful response",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/User"
                  }
                }
              }
            }
          }
        }
      },
      "/users/{id}": {
        "get": {
          "tags": [
            "users"
          ],
          "summary": "Get a user by ID",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "string"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Successful response",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/User"
                  }
                }
              }
            }
          }
        },
        "put": {
          "tags": [
            "users"
          ],
          "summary": "Update a user by ID",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "string"
              }
            }
          ],
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserInput"
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Successful response",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/User"
                  }
                }
              }
            }
          }
        },
        "delete": {
          "tags": [
            "users"
          ],
          "summary": "Delete a user by ID",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "string"
              }
            }
          ],
          "responses": {
            "204": {
              "description": "Successful response"
            }
          }
        }
      }
    },
    "components": {
      "schemas": {
...
        "Starship": {
          "type": "object",
          "properties": {
            "_id": {
              "type": "string"
            },
            "field": {
              "type": "object",
              "properties": {
                "MGLT": {
                  "type": "string"
                },
                "starship_class": {
                  "type": "string"
                },
                "hyperdrive_rating": {
                  "type": "string"
                },
                "pilots": {
                  "type": "array",
                  "items": {
                    "type": "integer"
                  }
                },
                "created": {
                  "type": "string",
                  "format": "date-time"
                },
                "edited": {
                  "type": "string",
                  "format": "date-time"
                }
              }
            },
            "model": {
              "type": "string"
            },
            "pk": {
              "type": "integer"
            }
          }
        },
        "StarshipInput": {
          "type": "object",
          "properties": {
            "MGLT": {
              "type": "string"
            },
            "starship_class": {
              "type": "string"
            },
            "hyperdrive_rating": {
              "type": "string"
            },
            "pilots": {
              "type": "array",
              "items": {
                "type": "integer"
              }
            }
          },
          "required": [
            "MGLT",
            "starship_class",
            "hyperdrive_rating",
            "pilots"
          ]
        },
        "Transport": {
          "type": "object",
          "properties": {
            "_id": {
              "type": "string"
            },
            "field": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string"
                },
                "cargo_capacity": {
                  "type": "string"
                },
                "consumables": {
                  "type": "string"
                },
                "passengers": {
                  "type": "string"
                },
                "max_atmosphering_speed": {
                  "type": "string"
                },
                "crew": {
                  "type": "string"
                },
                "length": {
                  "type": "string"
                },
                "model": {
                  "type": "string"
                },
                "cost_in_credits": {
                  "type": "string"
                },
                "manufacturer": {
                  "type": "string"
                },
                "created": {
                  "type": "string",
                  "format": "date-time"
                },
                "edited": {
                  "type": "string",
                  "format": "date-time"
                }
              }
            },
            "model": {
              "type": "string"
            },
            "pk": {
              "type": "integer"
            }
          }
        },
        "TransportInput": {
          "type": "object",
          "properties": {
            "name": {
              "type": "string"
            },
            "cargo_capacity": {
              "type": "string"
            },
            "consumables": {
              "type": "string"
            },
            "passengers": {
              "type": "string"
            },
            "max_atmosphering_speed": {
              "type": "string"
            },
            "crew": {
              "type": "string"
            },
            "length": {
              "type": "string"
            },
            "model": {
              "type": "string"
            },
            "cost_in_credits": {
              "type": "string"
            },
            "manufacturer": {
              "type": "string"
            }
          },
          "required": [
            "name",
            "cargo_capacity",
            "consumables",
            "passengers",
            "max_atmosphering_speed",
            "crew",
            "length",
            "model",
            "cost_in_credits",
            "manufacturer"
          ]
        },
        "Vehicle": {
          "type": "object",
          "properties": {
            "_id": {
              "type": "string"
            },
            "field": {
              "type": "object",
              "properties": {
                "vehicle_class": {
                  "type": "string"
                },
                "pilots": {
                  "type": "array",
                  "items": {
                    "type": "integer"
                  }
                },
                "created": {
                  "type": "string",
                  "format": "date-time"
                },
                "edited": {
                  "type": "string",
                  "format": "date-time"
                }
              }
            },
            "model": {
              "type": "string"
            },
            "pk": {
              "type": "integer"
            }
          }
        },
        "VehicleInput": {
          "type": "object",
          "properties": {
            "vehicle_class": {
              "type": "string"
            },
            "pilots": {
              "type": "array",
              "items": {
                "type": "integer"
              }
            }
          },
          "required": [
            "vehicle_class",
            "pilots"
          ]
        },
        "User": {
          "type": "object",
          "properties": {
            "email": {
              "type": "string"
            },
            "password": {
              "type": "string"
            }
          }
        },
        "UserInput": {
          "type": "object",
          "properties": {
            "email": {
              "type": "string"
            },
            "password": {
              "type": "string"
            }
          },
          "required": [
            "email",
            "password"
          ]
        }
      }
    }
  }